% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorbidity.asso.R
\name{comorbidity.asso}
\alias{comorbidity.asso}
\title{Generate comorbiity association result file}
\usage{
comorbidity.asso(
  res.file,
  cormorbidity.file,
  population = "all",
  covars = c("sex", "age", "bmi"),
  phe.name,
  output = NULL
)
}
\arguments{
\item{res.file}{Result summary file generated from COVID19.susceptibility, COVID19.severity or COVID19.mortality.}

\item{cormorbidity.file}{Comorbidity summary file generated from comorbidity.summary.}

\item{population}{Choose population from "all", white", "black", "asian", "mixed", or "other".}

\item{covars}{Selected covariates names. By default, covars=c("sex","age","bmi"), covariates are sex age and BMI.}

\item{phe.name}{Phenotype name.}

\item{output}{Name of comorbidity association test result file to be outputted.}
}
\value{
Outputs a comorbidity association test result file with OR, 95\% CI and p-value.
}
\description{
Generate comorbiity association result file
}
\examples{
comorbidity.asso(res.file, cormorbidity.file, population="white", covars=c("sex","age","bmi","SES","smoke","inAgedCare"), phe.name="severe.lev2", output = "lev2_bf.csv")
}
\keyword{association}
\keyword{comorbidity,}
\keyword{test.}
