% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COVID19.mortality.R
\name{COVID19.mortality}
\alias{COVID19.mortality}
\title{Combine COVID-19 mortality phenotype and covariates}
\usage{
COVID19.mortality(
  res.eng,
  res.wal = NULL,
  res.sco = NULL,
  death.file,
  death.cause.file,
  cov.file,
  Date = NULL,
  out.name = NULL
)
}
\arguments{
\item{res.eng}{Latest covid result file/files for England.}

\item{res.wal}{Latest covid result file/files for Wales. Only available for downloads after April 2021.}

\item{res.sco}{Latest covid result file/files for Scotland. Only available for downloads after April 2021.}

\item{death.file}{Latest death register file.}

\item{death.cause.file}{Latest death cause file.}

\item{cov.file}{Covariate file generated using risk.factor function.}

\item{Date}{Date, ddmmyyyy, select the results until a certain date. By default, Date = NULL, the latest death register date.}

\item{out.name}{Name of mortality file to be outputted. By default, out.name = NULL, “mortality.cov.txt”.}
}
\value{
Outputs a mortality file with phenotype and covariates.
}
\description{
Combine COVID-19 mortality phenotype and covariates
}
\examples{
mortality <- COVID19.mortality(res.eng=covid_example("sim_result_england.txt.gz"), 
death.file=covid_example("sim_death.txt.gz"), 
death.cause.file=covid_example("sim_death_cause.txt.gz"), 
cov.file=covid_example("covariate.txt"),
out.name=paste0(covid_example("results"),"/mortality"))

}
\keyword{covariates,}
\keyword{mortality,}
\keyword{phenotype}
