% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COVID19.susceptibility.R
\name{COVID19.susceptibility}
\alias{COVID19.susceptibility}
\title{Combine COVID-19 "susceptibility" phenotype and covariates}
\usage{
COVID19.susceptibility(
  res.eng,
  res.wal = NULL,
  res.sco = NULL,
  cov.file,
  Date = NULL,
  out.name = NULL
)
}
\arguments{
\item{res.eng}{Latest covid result file/files for England.}

\item{res.wal}{Latest covid result file/files for Wales. Only available for downloads after April 2021.}

\item{res.sco}{Latest covid result file/files for Scotland. Only available for downloads after April 2021.}

\item{cov.file}{Covariate file generated using risk.factor function.}

\item{Date}{Date, ddmmyyyy, select the results until a certain date. By default, Date = NULL, the latest testing date.}

\item{out.name}{Name of susceptibility file to be outputted. By default, out.name = NULL, “susceptibility.tested.txt” and “susceptibility.population.txt”.}
}
\value{
Outputs susceptibility files with phenotypes and covariates.
}
\description{
Combine COVID-19 "susceptibility" phenotype and covariates
}
\keyword{covariates}
\keyword{phenotype,}
\keyword{susceptibility,}
