% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_example.R
\name{covid_example}
\alias{covid_example}
\title{Provide working directory for UKB.COVID19 example files}
\usage{
covid_example(path)
}
\arguments{
\item{path}{path to file}
}
\value{
Outputs the working directory for UKB.COVID19 example files.
}
\description{
Provide working directory for UKB.COVID19 example files
}
\examples{
covid_example('results/covariate.txt')

}
