% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.reform.R
\name{data.reform}
\alias{data.reform}
\title{Reform variables}
\usage{
data.reform(res, type)
}
\arguments{
\item{res}{Merged data of phenotype from susceptibility.summary, severity.summary, mortality.summary or comorbidity.summary and covariates from covariate file.}

\item{type}{Data type: susceptibility, severity, mortality or comorbidity.}
}
\value{
Reformed data for association tests using logistic regression models.
}
\description{
Reform variables
}
\keyword{reform}
