% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_reform.R
\name{data_reform}
\alias{data_reform}
\title{Reform variables}
\usage{
data_reform(res, type)
}
\arguments{
\item{res}{Merged data of phenotype from makePhenotypes or comorbidity_summary and covariates from risk_factor.}

\item{type}{Data type: susceptibility, severity, mortality or comorbidity.}
}
\value{
Reformed data for association tests using logistic regression models.
}
\description{
Reform variables
}
\keyword{reform}
