% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asso.test.R
\name{log.cov}
\alias{log.cov}
\title{Perform association tests between phenotype and covariates}
\usage{
\method{log}{cov}(data, phe.name, cov.name = c("sex", "age", "bmi"), asso.output = NULL)
}
\arguments{
\item{data}{Data generated from COVID19.susceptibility, COVID19.severity or COVID19.mortality.}

\item{phe.name}{Phenotype name in the data.}

\item{cov.name}{Selected covariate names in the data. By default, cov.name=c("sex","age","bmi").}

\item{asso.output}{Name of association test result file to be outputted. By default, asso.output=NULL, it returns results but doesn't generate any files.}
}
\value{
Outputs association test results with OR, 95\% CI, and p-value.
}
\description{
This function formats and outputs a covariate file, used for input for other functions.
}
\examples{
log.cov(data=res$tested, phe.name="pos.neg", cov.name=c("sex","age","bmi"))
}
\keyword{association}
\keyword{test}
