% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality.summary.R
\name{mortality.summary}
\alias{mortality.summary}
\title{Create a phenotype for COVID-19 mortality}
\usage{
mortality.summary(
  res.eng,
  res.wal = NULL,
  res.sco = NULL,
  death.file,
  death.cause.file,
  Date = NULL,
  out.name = NULL
)
}
\arguments{
\item{res.eng}{Latest covid result file/files for England.}

\item{res.wal}{Latest covid result file/files for Wales. Only available for downloads after April 2021.}

\item{res.sco}{Latest covid result file/files for Scotland. Only available for downloads after April 2021.}

\item{death.file}{Latest death register file.}

\item{death.cause.file}{Latest death cause file.}

\item{Date}{Date, ddmmyyyy, select the results until a certain date. By default, Date = NULL, the latest death register date.}

\item{out.name}{Name of mortality file to be outputted. By default, out.name = NULL, “mortality_{Date}.txt”.}
}
\value{
Outputs a mortality summary file with mortality phenotype.
}
\description{
Create a phenotype for COVID-19 mortality
}
\keyword{mortality,}
\keyword{phenotype}
