% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.factor.R
\name{risk.factor}
\alias{risk.factor}
\title{Generate covariate file}
\usage{
risk.factor(
  ukb.data,
  ABO.data,
  hesin.file,
  res.eng,
  res.wal = NULL,
  res.sco = NULL,
  fields = NULL,
  field.names = NULL,
  out.file = NULL
)
}
\arguments{
\item{ukb.data}{tab delimited UK Biobank phenotype file.}

\item{ABO.data}{Latest yyyymmdd_covid19_misc.txt file.}

\item{hesin.file}{Latest yyyymmdd_hesin.txt file.}

\item{res.eng}{Latest covid result file/files for England.}

\item{res.wal}{Latest covid result file/files for Wales. Only available for downloads after April 2021.}

\item{res.sco}{Latest covid result file/files for Scotland. Only available for downloads after April 2021.}

\item{fields}{User specified field codes from ukb.data file.}

\item{field.names}{User specified field names.}

\item{out.file}{Name of covariate file to be outputted. By default, out.file = NULL, “covariate.txt”.}
}
\value{
Outputs a covariate file, used for input for other functions. Automatically returns sex, age at birthday in 2020, SES, self-reported ethnicity, most recently reported BMI, most recently reported pack-years, whether they reside in aged care (based on hospital admissions data, and covid test data) and blood type. Function also allows user to specify fields of interest (field codes, provided by UK Biobank), and allows the users to specify more intuitive names, for selected fields.
}
\description{
This function formats and outputs a covariate file, used for input for other functions.
}
\examples{
risk.factor(ukb.data=covid_example("sim_ukb.tab.gz"),
ABO.data=covid_example("sim_covid19_misc.txt.gz"),
hesin.file=covid_example("sim_hesin.txt.gz"),
res.eng=covid_example("sim_result_england.txt.gz"),
out.file=paste0(covid_example("results"),"/covariate"))

}
