% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/severity.R
\name{severity.summary}
\alias{severity.summary}
\title{Create phenotypes for COVID-19 severity}
\usage{
severity.summary(
  res.eng,
  res.wal = NULL,
  res.sco = NULL,
  death.file,
  death.cause.file,
  hesin.file,
  hesin_diag.file,
  hesin_oper.file,
  hesin_critical.file,
  code.file = "./data/coding240.tsv",
  Date = NULL,
  out.name = NULL
)
}
\arguments{
\item{res.eng}{Latest covid result file/files for England.}

\item{res.wal}{Latest covid result file/files for Wales. Only available for downloads after April 2021.}

\item{res.sco}{Latest covid result file/files for Scotland. Only available for downloads after April 2021.}

\item{death.file}{Latest death register file.}

\item{death.cause.file}{Latest death cause file.}

\item{hesin.file}{Latest hospital inpatient master file.}

\item{hesin_diag.file}{Latest hospital inpatient diagnosis file.}

\item{hesin_oper.file}{Latest hospital inpatient operation file.}

\item{hesin_critical.file}{Latest hospital inpatient critical care file.}

\item{code.file}{The operation code file, which is included in the package, and also can be download from https://github.com/bahlolab/UKB.COVID19/blob/main/data/coding240.tsv.}

\item{Date}{Date, ddmmyyyy, select the results until a certain date. By default, Date = NULL, the latest hospitalization date.}

\item{out.name}{Name of severity files to be outputted.}

\item{cov.file}{Covariate file generated using risk.factor function.}
}
\value{
Outputs a severity file with phenotypes for hospitalization, critical care and advanced critical care and covariates.
}
\description{
Create phenotypes for COVID-19 severity
}
\examples{
severity <- COVID19.severity(res.eng, res.wal, res.sco, death.file, death.cause.file, hesin.file, hesin_diag.file, hesin_oper.file, hesin_critical.file, cov.file, code.file)
}
\keyword{care,}
\keyword{critical}
\keyword{hospitalization,}
\keyword{phenotype}
\keyword{severity,}
