% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susceptibility.summary.R
\name{susceptibility.summary}
\alias{susceptibility.summary}
\title{Create phenotypes for COVID-19 "susceptibility"}
\usage{
susceptibility.summary(
  res.eng,
  res.wal = NULL,
  res.sco = NULL,
  Date = NULL,
  out.name = NULL
)
}
\arguments{
\item{res.eng}{Latest covid result file/files for England.}

\item{res.wal}{Latest covid result file/files for Wales. Only available for downloads after April 2021.}

\item{res.sco}{Latest covid result file/files for Scotland. Only available for downloads after April 2021.}

\item{Date}{Date, ddmmyyyy, select the results until a certain date. By default, Date = NULL, the latest testing date.}

\item{out.name}{Name of susceptibility file to be outputted. By default, out.name = NULL, “result_{Date}.txt”.}
}
\value{
Outputs susceptibility files: 1. positive vs negative; 2. positive vs population.
}
\description{
Create phenotypes for COVID-19 "susceptibility"
}
\keyword{phenotype}
\keyword{susceptibility,}
