% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorbidity.asso.R
\name{comorbidity.asso}
\alias{comorbidity.asso}
\title{Generate comorbidity association result file}
\usage{
comorbidity.asso(
  res.file,
  cormorbidity.file,
  population = "all",
  covars = c("sex", "age", "bmi"),
  phe.name,
  output = NULL
)
}
\arguments{
\item{res.file}{Result summary file generated from COVID19.susceptibility, COVID19.severity or COVID19.mortality.}

\item{cormorbidity.file}{Comorbidity summary file generated from comorbidity.summary.}

\item{population}{Choose self-report population/ethnic background group from "all", white", "black", "asian", "mixed", or "other". By default, population="all", include all ethnic groups.}

\item{covars}{Selected covariates names. By default, covars=c("sex","age","bmi"), covariates are sex age and BMI.}

\item{phe.name}{Phenotype name.}

\item{output}{Name of comorbidity association test result file to be outputted. By default, output=NULL, it is {population}_{phenotype name}_comorbidity_asso.csv.}
}
\value{
Outputs a comorbidity association test result file with OR, 95\% CI and p-value.
}
\description{
Association tests between each co-morbidity and given phenotype (susceptibility, mortality or severity) with the adjustment of covariates.
}
\examples{
\dontrun{
comorbidity.asso(res.file=covid_example("susceptibility.tested.txt"), 
cormorbidity.file=covid_example("comorbidity_1995-07-12_2020-12-31.RData"), 
population="white", 
covars=c("sex","age","bmi","SES","smoke","inAgedCare"), 
phe.name="pos.neg", 
output = "cormorb_susceptibility_asso.csv")
}

}
\keyword{association}
\keyword{comorbidity,}
\keyword{test.}
