% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comorbidity.summary.R
\name{comorbidity.summary}
\alias{comorbidity.summary}
\title{Create comorbidity summary file}
\usage{
comorbidity.summary(
  hesin.file,
  hesin_diag.file,
  cov.file,
  primary = FALSE,
  ICD10.file,
  Date.start = NULL,
  Date.end = NULL
)
}
\arguments{
\item{hesin.file}{Latest hospital inpatient master file.}

\item{hesin_diag.file}{Latest hospital inpatient diagnosis file.}

\item{cov.file}{Covariate file generated using risk.factor function.}

\item{primary}{TRUE: include primary diagnosis only; FALSE: include all diagnoses.}

\item{ICD10.file}{The ICD10 code file, which is included in the package, and also can be download from https://github.com/bahlolab/UKB.COVID19/blob/main/data/ICD10.coding19.tsv.}

\item{Date.start}{Date, ddmmyyyy, select the start date of hospital inpatient record period.}

\item{Date.end}{Date, ddmmyyyy, select the end date of hospital inpatient record period.}
}
\value{
Outputs comorbidity summary file, named comorbidity_<Date.start>_<Date.end>.RData, including phenotype, non-genetic risk factors and all comorbidities, which will be used in the comorbidity association tests.
}
\description{
summarises disease history records of each individual from the hospital inpatient diagnosis data.
}
\examples{
comorbidity.summary(hesin.file, hesin_diag.file, cov.file, ICD10.file, primary = F, Date.end="16/03/2020")
}
\keyword{comorbidity}
